requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
grpcio==1.65.4
grpcio-tools==1.65.4
idna==3.7
protobuf==5.27.3
requests==2.32.3
urllib3==2.2.2
